+++ lib/src/chart.dart
@@ -6,14 +6,8 @@ import 'package:flutter/foundation.dart';
 
 import 'chart_controller.dart';
 import 'chart_render.dart';
-import 'components/background/background.dart';
-import 'components/crosshair/crosshair.dart';
-import 'components/splitter/splitter.dart';
-import 'components/viewport_h.dart';
-import 'components/viewport_h_scaler.dart';
-import 'components/viewport_v.dart';
+import 'components/components.dart';
 import 'data/data_source.dart';
-import 'components/panel/panel.dart';
 import 'theme/theme.dart';
 import 'values/value.dart';
 
@@ -111,10 +105,12 @@ class GChart extends ChangeNotifier {
   /// The pre-render callback which is called right before rendering.
   ///
   /// This gives a chance to modify the chart before rendering.
-  final void Function(GChart chart, Size size)? preRender;
+  final void Function(GChart chart, Canvas canvas, Rect area)? preRender;
 
   /// The post-render callback which is called right after rendering finished.
-  final void Function(GChart chart, Size size)? postRender;
+  ///
+  /// still it is able to draw something additional on the canvas.
+  final void Function(GChart chart, Canvas canvas, Rect area)? postRender;
 
   final _debounceHelper = DebounceHelper(milliseconds: 500);
 
@@ -211,14 +207,25 @@ class GChart extends ChangeNotifier {
 
   /// Paint the chart on the canvas.
   void paint(Canvas canvas, Size size) {
-    if (kDebugMode && _paintCount.value % 100 == 0) {
-      // ignore: avoid_print
-      print("paintCount = ${_paintCount.value}");
+    if (kDebugMode) {
+      _paintCount.value += 1;
+      if (_paintCount.value % 100 == 0) {
+        // ignore: avoid_print
+        print("paintCount = ${_paintCount.value}");
+      }
     }
-    _paintCount.value += 1;
-    preRender?.call(this, area.size);
+    preRender?.call(this, canvas, area);
     render.render(canvas: canvas, chart: this);
-    postRender?.call(this, size);
+    postRender?.call(this, canvas, area);
+  }
+
+  /// Save current chart as an image.
+  Future<Image> saveAsImage() async {
+    final recorder = PictureRecorder();
+    final canvas = Canvas(recorder, area);
+    render.render(canvas: canvas, chart: this);
+    final picture = recorder.endRecording();
+    return picture.toImage(size.width.floor(), size.height.floor());
   }
 
   /// Resize the chart view area.
@@ -336,6 +343,10 @@ class GChart extends ChangeNotifier {
         }
       }
     }
+  }
+
+  (GPanel, GGraph)? hitTestGraph({required Offset position}) {
+    return controller.hitTestGraph(position: position);
   }
 
   void _pointViewPortChanged() {